refactor : request header 객체를 따로 생성
header를 map에서 객체로 변경하였다. 해당 객체는 header에서 필요한 정보인 sid 값을 멤버변수로 가지고 있다. 추후 header에서 또 이런 식으로 필요한 정보들이 생기면 추가가 가능할 것이다.
